# packages/api_client_dart/openapi.json
{
  "components": {
    "schemas": {
      "AttributesOut": {
        "properties": {
          "colours": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Colours",
            "type": "array"
          },
          "craft_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Craft Id"
          },
          "dimensions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimensions"
          },
          "hours_to_make": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hours To Make"
          },
          "materials": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Materials",
            "type": "array"
          },
          "missing": {
            "default": [],
            "description": "Fields still unfilled, in the order the app should ask about them aloud.",
            "items": {
              "type": "string"
            },
            "title": "Missing",
            "type": "array"
          },
          "product_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Type"
          },
          "quantity_available": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity Available"
          },
          "technique": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technique"
          }
        },
        "title": "AttributesOut",
        "type": "object"
      },
      "CatalogOut": {
        "properties": {
          "attributes": {
            "$ref": "#/components/schemas/AttributesOut"
          },
          "descriptions": {
            "items": {
              "$ref": "#/components/schemas/DescriptionOut"
            },
            "title": "Descriptions",
            "type": "array"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "transcript": {
            "title": "Transcript",
            "type": "string"
          }
        },
        "required": [
          "transcript",
          "language",
          "attributes",
          "descriptions"
        ],
        "title": "CatalogOut",
        "type": "object"
      },
      "CraftOut": {
        "properties": {
          "cluster": {
            "title": "Cluster",
            "type": "string"
          },
          "collection": {
            "title": "Collection",
            "type": "string"
          },
          "gi_tagged": {
            "title": "Gi Tagged",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "collection",
          "state",
          "cluster",
          "gi_tagged",
          "unit"
        ],
        "title": "CraftOut",
        "type": "object"
      },
      "DescriptionOut": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "title": "Keywords",
            "type": "array"
          },
          "locale": {
            "title": "Locale",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "locale",
          "title",
          "body",
          "keywords"
        ],
        "title": "DescriptionOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PriceIn": {
        "properties": {
          "craft_id": {
            "title": "Craft Id",
            "type": "string"
          },
          "finish_score": {
            "default": 0.5,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Finish Score",
            "type": "number"
          },
          "hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Hours"
          },
          "material_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Material Cost"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "craft_id",
          "state",
          "material_cost",
          "hours"
        ],
        "title": "PriceIn",
        "type": "object"
      },
      "PriceOut": {
        "properties": {
          "band_high": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Band High",
            "type": "string"
          },
          "band_low": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Band Low",
            "type": "string"
          },
          "currency": {
            "default": "INR",
            "title": "Currency",
            "type": "string"
          },
          "floor": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Floor",
            "type": "string"
          },
          "labour_cost": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Labour Cost",
            "type": "string"
          },
          "material_cost": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Material Cost",
            "type": "string"
          },
          "overhead": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Overhead",
            "type": "string"
          },
          "position": {
            "description": "'within_band' or 'floor_above_market'. When floor_above_market, the app must show the below-market state explicitly (Palette.belowFloor plus icon plus speech) -- never just the number.",
            "title": "Position",
            "type": "string"
          },
          "rationale": {
            "description": "Plain language, for text-to-speech. Read this aloud.",
            "title": "Rationale",
            "type": "string"
          },
          "stretch": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Stretch",
            "type": "string"
          },
          "suggested": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Suggested",
            "type": "string"
          }
        },
        "required": [
          "floor",
          "suggested",
          "stretch",
          "band_low",
          "band_high",
          "position",
          "rationale",
          "material_cost",
          "labour_cost",
          "overhead"
        ],
        "title": "PriceOut",
        "type": "object"
      },
      "TranscribeIn": {
        "properties": {
          "audio_base64": {
            "title": "Audio Base64",
            "type": "string"
          },
          "language": {
            "examples": [
              "te-IN",
              "hi-IN"
            ],
            "title": "Language",
            "type": "string"
          }
        },
        "required": [
          "language",
          "audio_base64"
        ],
        "title": "TranscribeIn",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Market on-ramp for marginalised artisans. SIH 2026, PS 26090 (Ministry of Social Justice and Empowerment).",
    "title": "Karigar API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/catalog/from-voice": {
      "post": {
        "description": "Voice note in, structured listing out.\n\nThe transcript is never handed to the writer as free text -- it is\nextracted into a typed schema first, so missing fields become spoken\nfollow-up questions rather than model invention.",
        "operationId": "catalogFromVoice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscribeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Catalog From Voice",
        "tags": [
          "catalog"
        ]
      }
    },
    "/v1/crafts": {
      "get": {
        "operationId": "listCrafts",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CraftOut"
                  },
                  "title": "Response Listcrafts",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Crafts",
        "tags": [
          "catalog"
        ]
      }
    },
    "/v1/health": {
      "get": {
        "operationId": "health",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "ops"
        ]
      }
    },
    "/v1/pricing/quote": {
      "post": {
        "description": "Floor, band and suggestion -- never one opaque number (Bet 03).",
        "operationId": "quotePrice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Quote",
        "tags": [
          "pricing"
        ]
      }
    }
  }
}